+++ src/Views/Traits/Columns/IsSelectable.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Views\Traits\Columns;
+
+use Closure;
+use Rappasoft\LaravelLivewireTables\DataTableComponent;
+use Rappasoft\LaravelLivewireTables\Views\Column;
+
+trait IsSelectable
+{
+
+    protected bool $selectable = true;
+
+    protected bool $selected = true;
+
+
+    public function isSelectable(): bool
+    {
+        return $this->selectable === true;
+    }
+
+    public function isSelected(): bool
+    {
+        return $this->selected === true;
+    }
+
+    
+    public function excludeFromColumnSelect(): self
+    {
+        $this->selectable = false;
+
+        return $this;
+    }
+
+    public function deselected(): self
+    {
+        $this->selected = false;
+
+        return $this;
+    }
+
+}
