+++ src/Views/Traits/IsColumn.php
@@ -3,9 +3,10 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Traits;
 
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
-use Rappasoft\LaravelLivewireTables\Views\Traits\Columns\{IsCollapsible, IsSearchable, IsSortable};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Columns\{HasVisibility, IsCollapsible, IsSearchable, IsSelectable, IsSortable};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\ColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasAttributes,HasFooter,HasSecondaryHeader,HasView};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Columns\HasVisibility;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\{ColumnHelpers,RelationshipHelpers};
 
 trait IsColumn
@@ -15,11 +16,13 @@ trait IsColumn
         RelationshipHelpers,
         IsCollapsible,
         IsSearchable,
+        IsSelectable,
         IsSortable,
         HasAttributes,
         HasFooter,
         HasSecondaryHeader,
-        HasView;
+        HasView,
+        HasVisibility;
 
     protected ?DataTableComponent $component = null;
 
@@ -48,12 +51,6 @@ trait IsColumn
     protected bool $html = false;
 
     protected mixed $labelCallback = null;
-
-    protected bool $hidden = false;
-
-    protected bool $selectable = true;
-
-    protected bool $selected = true;
 
     protected bool $clickable = true;
